requirements.txt
alpaca-trade-api==3.1.1
numpy==1.24.3
pandas==2.0.1
python-dotenv==1.0.0
scikit-learn==1.2.2
loguru==0.7.0
Flask==3.0.2
Flask-SocketIO==5.3.6
